\name{SimulVals}
\alias{SimulVals}
\title{Pre-simulated Joint Binary Probabilities}
\description{
  This variable provides a pre-fabricated result from
  \code{\link{simul.commonprob}} such that it is normally not necessary
  to use this (time consuming) function, and is used by
  \code{\link{rmvbin}}.
}
\usage{
SimulVals
}
\references{Friedrich Leisch, Andreas Weingessel and Kurt Hornik
  (1998). On the generation of correlated artificial binary
  data. Working Paper Series, SFB ``Adaptive Information Systems and
  Modelling in Economics and Management Science'', Vienna University of
  Economics, \url{http://www.wu-wien.ac.at/am}
}
\author{Friedrich Leisch}
\seealso{
  \code{\link{simul.commonprob}}, \code{\link{rmvbin}}
}
\keyword{sysdata}
